requirements.txt
PyNaCl==1.4.0
aiohttp==3.8.0
nextcord==2.0.0a3
pyyaml==6.0
